Distributed twobody tests into modules
